2nd commit - updated request function & fixed some things based on your comment
