+++ src/App.js
@@ -4,21 +4,19 @@ import "./App.css";
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 
 import HomePage from "./pages/HomePage";
-import NavBar from "./components/NavBar/Nav";
+import NavBar from "./components/Nav";
 import LogInPage from "./pages/LogInPage";
 import SignUpPage from "./pages/SignUpPage/SignUpPage";
-import RequestPage from "./pages/RequestPage";
 
 function App() {
   return (
     <div className="App">
-      <NavBar />
       <Router>
+      <NavBar />
         <Switch>
           <Route exact path="/" component={HomePage} />
           <Route exact path="/login" component={LogInPage} />
           <Route exact path="/signup" component={SignUpPage} />
-          <Route exact path="/request" component={RequestPage} />
         </Switch>
       </Router>
     </div>
