--- src/components/NavBar/styles.css
@@ -1,5 +1,0 @@
-.NavBar{
-    display: flex;
-    justify-content: space-between;
-    width: 100%;
-}
