+++ src/pages/HomePage.js
@@ -3,6 +3,7 @@ import React from "react";
 
 
 const HomePage = () => {
+  console.log(localStorage.getItem("isLoggedIn"))
   return (
     <div>
       <h1> Home Page Goes Here </h1>
