--- src/pages/RequestPage.js
@@ -1,13 +1,0 @@
-import React from "react";
-
-
-
-const RequestPage = () => {
-
-
-    return (
-        <h1>Request page here</h1>
-    )
-};
-
-export default RequestPage;
