+++ src/pages/SignUpPage/signUp.styles.css
@@ -1,11 +1,11 @@
 .formSignUp {
-    border: 1px solid blue;
+    border: 1px solid #ef8172;
     border-radius: 10px;
     width:60%;
     padding: 3%;
     margin: auto;
     margin-top: 5%;
-    box-shadow: 5px 7px #456fa7;
+    box-shadow: 5px 7px #ef8172;
     display: flex;
     flex-direction: column;
     justify-content: space-around;
@@ -20,7 +20,3 @@
     margin: auto;
 }
 
-h1 {
-    color: #227aec;
-    margin-bottom: 2rem;
-}
